requirements.txt
Django==5.2.0
django-jazzmin==3.0.4
djangorestframework==3.17.1
drf-yasg==1.21.15
gunicorn==23.0.0
pillow==12.2.0
qrcode==8.2
whitenoise==6.11.0
